requirements.txt
blinkpy>=0.23.0
python-telegram-bot>=20.0
python-dotenv>=0.19.0
aiohttp>=3.8.0 